cardの分離終わり
